+++ src/features/linter-provider.ts
@@ -10,14 +10,15 @@ import { Logger } from '../services/logging';
 import { GNULinter, GNUModernLinter, IntelLinter, LFortranLinter, NAGLinter } from '../lib/linters';
 import {
   EXTENSION_ID,
-  FortranDocumentSelector,
   resolveVariables,
   promptForMissingTool,
   isFreeForm,
   spawnAsPromise,
+  isFortran,
+  shellTask,
 } from '../lib/tools';
 import { arraysEqual } from '../lib/helper';
-import { RescanLint } from './commands';
+import { BuildDebug, BuildRun, RescanLint } from './commands';
 import { GlobPaths } from '../lib/glob-paths';
 
 export class LinterSettings {
@@ -164,7 +165,24 @@ export class FortranLintingProvider {
   public async activate(subscriptions: vscode.Disposable[]) {
     // Register Linter commands
     subscriptions.push(vscode.commands.registerCommand(RescanLint, this.rescanLinter, this));
-
+    subscriptions.push(
+      vscode.commands.registerTextEditorCommand(
+        BuildRun,
+        async (textEditor: vscode.TextEditor, edit: vscode.TextEditorEdit, ...args: any[]) => {
+          await this.buildAndRun(textEditor);
+        },
+        this
+      )
+    );
+    subscriptions.push(
+      vscode.commands.registerTextEditorCommand(
+        BuildDebug,
+        async (textEditor: vscode.TextEditor, edit: vscode.TextEditorEdit, ...args: any[]) => {
+          await this.buildAndDebug(textEditor);
+        },
+        this
+      )
+    );
     vscode.workspace.onDidOpenTextDocument(this.doLint, this, subscriptions);
     vscode.workspace.onDidCloseTextDocument(
       textDocument => {
@@ -194,12 +212,7 @@ export class FortranLintingProvider {
     // Only lint if a compiler is specified
     if (!this.settings.enabled) return;
     // Only lint Fortran (free, fixed) format files
-    if (
-      !FortranDocumentSelector().some(e => e.scheme === textDocument.uri.scheme) ||
-      !FortranDocumentSelector().some(e => e.language === textDocument.languageId)
-    ) {
-      return;
-    }
+    if (!isFortran(textDocument)) return;
 
     this.linter = this.getLinter(this.settings.compiler);
     const command = this.getLinterExecutable();
@@ -252,6 +265,45 @@ export class FortranLintingProvider {
     }
   }
 
+  private async buildAndRun(textEditor: vscode.TextEditor) {
+    return this.buildAndDebug(textEditor, false);
+  }
+
+  /**
+   * Compile and run the current file using the provided linter options.
+   * It has the ability to launch a Debug session or just run the executable.
+   * @param textEditor a text editor instance
+   * @param debug performing a debug build or not
+   */
+  private async buildAndDebug(textEditor: vscode.TextEditor, debug = true): Promise<void> {
+    const textDocument = textEditor.document;
+    this.linter = this.getLinter(this.settings.compiler);
+    const command = this.getLinterExecutable();
+    let argList = [...this.constructArgumentList(textDocument)];
+    // Remove mandatory linter args, used for mock compilation
+    argList = argList.filter(arg => !this.linter.args.includes(arg));
+    if (debug) argList.push('-g'); // add debug symbols flag, same for all compilers
+    try {
+      await shellTask(command, argList, 'Build Fortran file');
+      const folder: vscode.WorkspaceFolder = vscode.workspace.getWorkspaceFolder(
+        textEditor.document.uri
+      );
+      const selectedConfig: vscode.DebugConfiguration = {
+        name: `${debug ? 'Debug' : 'Run'} Fortran file`,
+        // This relies on the C/C++ debug adapters
+        type: process.platform === 'win32' ? 'cppvsdbg' : 'cppdbg',
+        request: 'launch',
+        program: `${textDocument.fileName}.o`,
+        cwd: folder.uri.fsPath,
+      };
+      await vscode.debug.startDebugging(folder, selectedConfig, { noDebug: debug });
+      return;
+    } catch (err) {
+      this.logger.error(`[build] Compiling ${textDocument.fileName} failed:`, err);
+      console.error(`ERROR: ${err}`);
+    }
+  }
+
   private getLinter(compiler: string): GNULinter | GNUModernLinter | IntelLinter | NAGLinter {
     switch (compiler) {
       case 'gfortran':
@@ -276,8 +328,8 @@ export class FortranLintingProvider {
     this.logger.debug(`[lint] glob paths:`, this.pathCache.get(opt).globs);
     this.logger.debug(`[lint] resolved paths:`, this.pathCache.get(opt).paths);
 
-    const extensionIndex = textDocument.fileName.lastIndexOf('.');
-    const fileNameWithoutExtension = textDocument.fileName.substring(0, extensionIndex);
+    // const extensionIndex = textDocument.fileName.lastIndexOf('.');
+    // const fileNameWithoutExtension = textDocument.fileName.substring(0, extensionIndex);
     const fortranSource: string[] = this.settings.fyppEnabled
       ? ['-xf95', isFreeForm(textDocument) ? '-ffree-form' : '-ffixed-form', '-']
       : [textDocument.fileName];
@@ -286,7 +338,7 @@ export class FortranLintingProvider {
       ...args,
       ...this.getIncludeParams(includePaths), // include paths
       '-o',
-      `${fileNameWithoutExtension}.mod`,
+      `${textDocument.fileName}.o`,
       ...fortranSource,
     ];